backend/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Cache
redis==5.0.1
hiredis==2.2.3

# LLM SDK
anthropic==0.7.7
openai==1.3.7
google-generativeai==0.3.1
tiktoken==0.5.2

# Data processing
pandas==2.1.3
numpy==1.26.2
xlsxwriter==3.1.9
openpyxl==3.1.2

# Database drivers
clickhouse-driver==0.2.6
pymysql==1.1.0
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging
loguru==0.7.2

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Async tasks
celery==5.3.4
flower==2.0.1

# File processing
python-magic==0.4.27
chardet==5.2.0

# Lark SDK
lark-oapi==1.2.8

# Data validation
jsonschema==4.20.0
marshmallow==3.20.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx-mock==0.11.0

# Code quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.2

# Monitoring
prometheus-client==0.19.0
